Complete day 2, improve template, add parser

+++ src/day1/mod.rs
@@ -2,7 +2,7 @@ use itertools::izip;
 
 use crate::common;
 
-pub fn solver(star: u8) -> fn(String) -> usize {
+pub fn solver(star: u8) -> fn(String) -> i32 {
     match star {
         1 => star1,
         2 => star2,
@@ -10,16 +10,16 @@ pub fn solver(star: u8) -> fn(String) -> usize {
     }
 }
 
-fn star1(input: String) -> usize {
+fn star1(input: String) -> i32 {
     let depths = common::read_integers(&input);
 
     izip!(&depths, &depths[1..])
         .map(|(d1, d2)| d2 - d1)
         .filter(|d| d > &0)
-        .count()
+        .count() as i32
 }
 
-fn star2(input: String) -> usize {
+fn star2(input: String) -> i32 {
     let depths = common::read_integers(&input);
 
     let sums: Vec<_> = izip!(&depths, &depths[1..], &depths[2..])
@@ -29,5 +29,5 @@ fn star2(input: String) -> usize {
     izip!(&sums, &sums[1..])
         .map(|(d1, d2)| d2 - d1)
         .filter(|d| d > &0)
-        .count()
+        .count() as i32
 }
